requirement.txt
Babel==2.9.1
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
facebook-sdk==3.1.0
future==0.18.2
idna==3.3
importlib-metadata==4.11.1
Jinja2==3.0.3
MarkupSafe==2.1.0
oauthlib==3.2.0
Pillow==9.0.1
pycparser==2.21
pyrsistent==0.18.1
python-dateutil==2.8.2
python-editor==1.0.4
python-twitter==3.5
pytz==2021.3
qrcode==7.3.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
speaklater==1.3
urllib3==1.26.8
zipp==3.7.0
